[step5] refactor String member to Name class member
