+++ src/test/java/race/CarTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import race.domain.Car;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
